Improve ESLint and Code Quality
This commit addresses a number of ESLint violations and improves the overall code quality of the repository.

Key changes include:
- Updated the ESLint configuration to be more strict, turning several warnings into errors.
- Fixed all existing ESLint warnings, including:
  - Missing dependencies in useEffect hooks.
  - Replaced all instances of `any` with more specific TypeScript types.
  - Fixed issues with empty object types (`{}`).
  - Resolved a `react-refresh` warning by moving a large number of type definitions into a dedicated `src/types/hover.ts` file for better organization.
- Resolved a React warning in the test suite related to incorrect HTML nesting.
