+++ eslint.config.js
@@ -23,9 +23,9 @@ export default tseslint.config(
         'warn',
         { allowConstantExport: true },
       ],
-        '@typescript-eslint/no-explicit-any': 'warn',
-        "@typescript-eslint/no-empty-object-type": "warn",
-
+        '@typescript-eslint/no-explicit-any': 'error',
+        "@typescript-eslint/no-empty-object-type": "error",
+        "react-hooks/exhaustive-deps": "error"
     },
   },
 )
